correction exec7

+++ exec7.c
@@ -3,8 +3,8 @@
 int main() {
     int distance_km = 100;
     int vitesse_kmh = 50;
-    int temps =(float) distance_km / vitesse_kmh; 
-    int minutes =(int) temps * 60;
+    float temps = distance_km / vitesse_kmh; 
+    int minutes = temps * 60;
     printf("minutes = %d\n", minutes); 
     return 0;
 }
